# api/swagger/swagger.yaml
swagger: "2.0"
info:
  version: "0.0.1"
  title: Pedas App
# during dev, should point to your local machine

#host: 
host: localhost:10010
...
basePath: /
...
schemes:
  # tip: remove http to make production-grade
  - http
  - https
# format of bodies a client can send (Content-Type)
consumes:
  - application/json
# format of the responses to the client (Accepts)
produces:
  - application/json
paths:
...
  /hosts:
    x-swagger-router-controller: host.controller

    get:
      description: get the host list
      parameters:
        - name: nombre
          in: query
          type: string
          required: false
        - name: direccion
          in: query
          type: string
          required: false
        - name: sort
          in: query
          type: string
          required: false
      operationId: getHost
      responses: 
        "200":
          description: success
          schema: 
            $ref: "#/definitions/GetHostListResponse"
        default:
          description: Error
          schema: 
            $ref: "#/definitions/ErrorResponse"
            
    post:
      description: Add a new host to the list
      parameters:
        - in: body
          name: host
          description: The host to create
          schema:
            $ref: "#/definitions/Host"
      operationId: createHost
      responses: 
        "201":
          description: success
          schema:
            $ref: "#/definitions/GetHostResponse"
        default:
          description: Error
          schema: 
            $ref: "#/definitions/ErrorResponse"
  /hosts/{id}:
    x-swagger-router-controller: host.controller
    
    get: 
      description: Get a host
      parameters:
        - name: id
          type: string
          in: path
          required: true
      operationId: getHostById
      responses:
        "200":
          description: success
          schema:
           $ref: "#/definitions/GetHostResponse"
        default:
          description: Error
          schema: 
            $ref: "#/definitions/ErrorResponse"
            
    put:
      description: Update a host
      parameters:
        - name: id
          description: Host id
          type: string
          in: path
          required: true
        - in: body
          name: host
          description: The host to update
          schema:
            $ref: "#/definitions/Host"
      operationId: updateHost
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/GetHostResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
    
    delete:
      description: Delete a host
      parameters:
        - name: id
          description: Host id
          type: string
          in: path 
          required: true
      operationId: deleteHost
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/GeneralResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /alcoholes:
    x-swagger-router-controller: alcohol.controller

    get:
      description: get the alcohol list
      parameters:
        - name: nombre
          in: query
          type: string
          required: false
        - name: tipo
          in: query
          type: string
          required: false
        - name: sort
          in: query
          type: string
          required: false
      operationId: getAlcohol
      responses: 
        "200":
          description: success
          schema: 
            $ref: "#/definitions/GetAlcoholListResponse"
        default:
          description: Error
          schema: 
            $ref: "#/definitions/ErrorResponse"
            
    post:
      description: Add a new alcohol to the list
      parameters:
        - in: body
          name: alcohol
          description: The alcohol to create
          schema:
            $ref: "#/definitions/Alcohol"
      operationId: createAlcohol
      responses: 
        "201":
          description: success
          schema:
            $ref: "#/definitions/GetAlcoholResponse"
        default:
          description: Error
          schema: 
            $ref: "#/definitions/ErrorResponse"
  /alcoholes/{id}:
    x-swagger-router-controller: alcohol.controller
    
    get: 
      description: Get an alcohol
      parameters:
        - name: id
          type: string
          in: path
          required: true
      operationId: getAlcoholById
      responses:
        "200":
          description: success
          schema:
           $ref: "#/definitions/GetAlcoholResponse"
        default:
          description: Error
          schema: 
            $ref: "#/definitions/ErrorResponse"
            
    put:
      description: Update an alcohol
      parameters:
        - name: id
          description: Alcohol id
          type: string
          in: path
          required: true
        - in: body
          name: alcohol
          description: The alcohol to update
          schema:
            $ref: "#/definitions/Alcohol"
      operationId: updateAlcohol
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/GetAlcoholResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
    
    delete:
      description: Delete an alcohol
      parameters:
        - name: id
          description: Alcohol id
          type: string
          in: path 
          required: true
      operationId: deleteAlcohol
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/GeneralResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
...
  /hostpedas:
    x-swagger-router-controller: hostpedas.controller

    get:
      description: get the hostpedas list
      parameters:
        - name: host
          in: query
          type: number
          required: false
        - name: peda
          in: query
          type: number
          required: false
        - name: calificacion
          in: query
          type: string
          required: false
        - name: sort
          in: query
          type: string
          required: false
      operationId: getHostpedas
      responses: 
        "200":
          description: success
          schema: 
            $ref: "#/definitions/GetHostPedasListResponse"
        default:
          description: Error
          schema: 
            $ref: "#/definitions/ErrorResponse"
            
    post:
      description: Add a new hostpedas to the list
      parameters:
        - in: body
          name: hostpedas
          description: The hostpedas to create
          schema:
            $ref: "#/definitions/HostPedas"
      operationId: createHostpedas
      responses: 
        "201":
          description: success
          schema:
            $ref: "#/definitions/GetHostPedasResponse"
        default:
          description: Error
          schema: 
            $ref: "#/definitions/ErrorResponse"
  /hostpedas/{id}:
    x-swagger-router-controller: hostpedas.controller
    
    get: 
      description: Get a hostpedas
      parameters:
        - name: id
          type: string
          in: path
          required: true
      operationId: getHostpedasById
      responses:
        "200":
          description: success
          schema:
           $ref: "#/definitions/GetHostPedasResponse"
        default:
          description: Error
          schema: 
            $ref: "#/definitions/ErrorResponse"
            
    put:
      description: Update a hostpedas
      parameters:
        - name: id
          description: Hostpedas id
          type: string
          in: path
          required: true
        - in: body
          name: hostpedas
          description: The hostpedas to update
          schema:
            $ref: "#/definitions/HostPedas"
      operationId: updateHostpedas
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/GetHostPedasResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
    
    delete:
      description: Delete a hostpedas
      parameters:
        - name: id
          description: Hostpedas id
          type: string
          in: path 
          required: true
      operationId: deleteHostpedas
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/GeneralResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
...
definitions:
...
  Host:
    type: object
    properties:
      nombre:
        type: string
        description: Name of the host
      direccion:
        type: string
        description: Address of the host
      image:
        type: string
        format: byte
    required:
      - nombre
      - direccion
  GetHostResponse:
    type: object
    properties:
      id:
        type: number
        description: identifier
      nombre:
        type: string
        description: Name of the host
      direccion:
        type: string
        description: Address of the host
      image:
        type: string
        format: byte
    required:
      - id
      - nombre
      - direccion
  GetHostListResponse:
    required: 
      - hosts
    properties: 
      hosts:
        type: array
        items:
          $ref: "#/definitions/GetHostResponse"
  Alcohol:
    type: object
    properties:
      nombre:
        type: string
        description: Name of the alcohol
      tipo:
        type: string
        description: Type of the alcohol
      image:
        type: string
        format: byte
    required:
      - nombre
      - tipo
  GetAlcoholResponse:
    type: object
    properties:
      id:
        type: number
        description: identifier
      nombre:
        type: string
        description: Name of the alcohol
      tipo:
        type: string
        description: Type of the alcohol
      image:
        type: string
        format: byte
    required:
      - id
      - nombre
      - tipo
  GetAlcoholListResponse:
    required: 
      - alcoholes
    properties: 
      alcoholes:
        type: array
        items:
          $ref: "#/definitions/GetAlcoholResponse"
...
  HostPedas:
    type: object
    properties:
      host:
        type: number
        description: host
      peda:
        type: number
        description: peda
      calificacion:
        type: string
        description: calificacion
      image:
        type: string
        format: byte
    required:
      - host
      - peda
      - calificacion
  GetHostPedasResponse:
    type: object
    properties:
      id:
        type: number
        description: identifier
      host:
        type: number
        description: host
      peda:
        type: number
        description: peda
      calificacion:
        type: string
        description: calificacion
      image:
        type: string
        format: byte
    required:
      - id
      - host
      - peda
      - calificacion
  GetHostPedasListResponse:
    required: 
      - hostpedas
    properties: 
      hostpedas:
        type: array
        items:
          $ref: "#/definitions/GetHostPedasResponse"
  GeneralResponse:
    type: object
    properties:
      success:
        type: number
        description: Returns 1 if successful
      description:
        type: string
        description: A short comment
    required:
      - success
      - description
  ErrorResponse:
    required:
      - message
    properties:
      code:
        type: number
      message:
        type: string
      description:
        type: string